Refactor bot
