+++ tool/tctl/common/resources/resource.go
@@ -33,6 +33,7 @@ import (
 // to the Handler format.
 func Handlers() map[string]Handler {
 	return map[string]Handler{
+		types.KindBot:        botHandler(),
 		types.KindRole:       roleHandler(),
 		types.KindUser:       userHandler(),
 		types.KindDatabase:   databaseHandler(),
